requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.12
django-filter==2.2.0
django-filters==0.2.1
django-localflavor==2.1
djangorestframework==3.10.3
djangorestframework-gis==0.14
drf-yasg==1.17.1
GDAL==2.4.2
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
psycopg2==2.8.5
pyparsing==2.4.7
python-stdnum==1.12
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
